Added atomic block to timer and input object, removed input objects (deprecated)
- Added atomic block to timer and input object (16 bit data on 8 bit MCU that gets increment inside an interrupt)
- Removed input object list (deprecated)
- Power switch made into a single function
- Audio-system task does not get blocked

+++ ZZ/code/libs/castimer/castimer.h
@@ -12,7 +12,7 @@ private:
 	unsigned short timer_value = 0;
 	
 public:
-    unsigned long vrednost();
+    unsigned short vrednost();
     void ponastavi();
 	void increment();
 	class_TIMER(Vozlisce <class_TIMER*> &timer_list);
